Template distribute entity data
